APB-51 Remove abstraction layer for event handlers and event manager
