+++ modEvtMgrImpl/src/org/aion/evtmgr/impl/handler/BlockHandler.java
@@ -26,17 +26,16 @@ package org.aion.evtmgr.impl.handler;
 
 
 import org.aion.evtmgr.IHandler;
-import org.aion.evtmgr.impl.abs.AbstractHandler;
+import org.aion.evtmgr.impl.abs.EventHandler;
 
 /**
  * @author jay
  *
  */
-public class BlockHandler extends AbstractHandler implements IHandler {
+public class BlockHandler extends EventHandler implements IHandler {
 
     // Default constructor to set name of the thread, simplifies troubleshooting
     public BlockHandler() {
-        super(TYPE.BLOCK0.getValue());
-        dispatcher.setName("BlkHdr");
+        super(TYPE.BLOCK0.getValue(), "BlkHdr");
     }
 }
