+++ modEvtMgrImpl/src/org/aion/evtmgr/impl/handler/ConsensusHandler.java
@@ -25,17 +25,16 @@
 package org.aion.evtmgr.impl.handler;
 
 import org.aion.evtmgr.IHandler;
-import org.aion.evtmgr.impl.abs.AbstractHandler;
+import org.aion.evtmgr.impl.abs.EventHandler;
 
 /**
  * @author jay
  *
  */
-public class ConsensusHandler extends AbstractHandler implements IHandler {
+public class ConsensusHandler extends EventHandler implements IHandler {
 
     // Default constructor to set name of the thread, simplifies troubleshooting
     public ConsensusHandler() {
-        super(TYPE.CONSENSUS.getValue());
-        dispatcher.setName("ConsHdr");
+        super(TYPE.CONSENSUS.getValue(), "ConsHdr");
     }
 }
